add polyfill to testing
