+++ jest.setup.ts
@@ -3,3 +3,6 @@ import { TextEncoder, TextDecoder } from 'util';
 
 global.TextEncoder = TextEncoder;
 global.TextDecoder = TextDecoder;
+
+// polyfill for testing
+import './src/polyfill';
