+++ package.json
@@ -55,7 +55,7 @@
     "@types/react-modal": "^3.13.1",
     "@types/react-router-dom": "^5.3.3",
     "crypto-browserify": "^3.12.0",
-    "esbuild-loader": "^2.19.0",
+    "esbuild-loader": "^3.1.0",
     "firebase-tools": "^10.2.2",
     "https-browserify": "^1.0.0",
     "patch-package": "^6.5.1",
@@ -64,7 +64,7 @@
     "sass": "^1.53.0",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
-    "typescript": "^4.2.4"
+    "typescript": "^5.1.6"
   },
   "scripts": {
     "start": "cross-env IMAGE_INLINE_SIZE_LIMIT=1000 GENERATE_SOURCEMAP=false react-app-rewired start",
