+++ src/index.tsx
@@ -1,8 +1,8 @@
-import 'polyfill';
 import * as Sentry from '@sentry/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ToastProvider } from 'components/Toasts/context';
-import { initClient } from 'polyfill';
+import { initClient } from 'libs/utils';
+import 'polyfill';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
